cycle-6/requirements.txt
cryptography
pypdf
colorama